Moved setup onboarding handoff to React
no issue

The setup completion and checklist initialization flow now belongs to the React admin app, so Ember only hands off to the React setup completion route and no longer owns setup/done or onboarding preference writes.
